refactor: extract processInitCommand
